Clean check in prepublishOnly
Runs clean check as a first step in prepublishOnly.
This ensures a clean directory before building the build content and
before the manual diffing with previously published tarball.

This could be replaced with the actual `git clean -fdx` command but
trying to be cautious about running a mass deletion command on behalf
of the user.